# docs/api/openapi.m3.yaml
openapi: 3.1.0
info:
  title: Zalo Shop M3 Inventory, Search, Pricing and Cart API
  version: 0.3.4-m3.6
  description: >-
    M3.1 冻结契约；作为 openapi.m1.yaml 与 openapi.m2.yaml 的向后兼容增量。
    商城只从受信 X-Store-Code 与认证主体解析，请求体不接受 store_id。
    M3.2-M3.5 已实施数据库基础、库存、搜索、促销、优惠券、可信报价与管理员预览；M3.6 已实施会员购物车与移动端接入。
    本契约不进入 M4 订单、支付、物流或售后。
servers:
  - url: http://localhost:3000
...
paths:
...
  /v1/members/me/coupons:
    get:
      tags: [Member coupons]
      summary: 读取当前会员在当前商城领取的优惠券
      description: M3 只返回领取与资格状态，不伪造订单核销。
      security: [{ StoreCode: [], MemberBearer: [] }]
      parameters:
        - $ref: '#/components/parameters/StoreCode'
        - $ref: '#/components/parameters/Cursor'
        - name: limit
          in: query
          schema: { type: integer, minimum: 1, maximum: 50, default: 20 }
        - name: status
          in: query
          schema: { type: string, enum: [CLAIMED, EXPIRED, DISABLED] }
      responses:
        '200':
          description: 当前商城当前会员优惠券分页
          content:
            { application/json: { schema: { $ref: '#/components/schemas/MemberCouponPage' } } }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
  /v1/members/me/coupons/by-code/{couponCode}:
    put:
      tags: [Member coupons]
      summary: 确保当前会员已领取当前商城优惠券
      description: >-
        按会员和优惠券天然幂等；重复 PUT 返回同一领取记录，不增加 claimed_count。
        M3 不创建 REDEEMED 状态或订单核销记录。
      security: [{ StoreCode: [], MemberBearer: [] }]
      parameters:
        - $ref: '#/components/parameters/StoreCode'
        - $ref: '#/components/parameters/CouponCode'
      responses:
        '200':
          description: 首次领取或既有领取记录
          content: { application/json: { schema: { $ref: '#/components/schemas/MemberCoupon' } } }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
        '429': { $ref: '#/components/responses/RateLimited' }
...
  /v1/cart:
    get:
      tags: [Cart]
      summary: 读取当前商城会员活动购物车并实时重算
      security: [{ StoreCode: [], MemberBearer: [] }]
      parameters:
        - $ref: '#/components/parameters/StoreCode'
        - $ref: '#/components/parameters/Locale'
      responses:
        '200':
          description: 活动购物车；无车时返回空活动车
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Cart' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
  /v1/cart/items/by-sku/{skuCode}:
    put:
      tags: [Cart]
      summary: 按同商城 SKU 设置购物车数量
      description: 相同请求设置相同最终数量，不执行累加；加入购物车不锁库存。
      security: [{ StoreCode: [], MemberBearer: [] }]
      parameters:
        - $ref: '#/components/parameters/StoreCode'
        - $ref: '#/components/parameters/SkuCode'
        - $ref: '#/components/parameters/Locale'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/SetCartItemRequest' }
      responses:
        '200':
          description: 已设置并重算
          content: { application/json: { schema: { $ref: '#/components/schemas/Cart' } } }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
  /v1/cart/items/{itemId}:
    patch:
      tags: [Cart]
      summary: 修改数量、选中状态或替换同商品 SKU
      security: [{ StoreCode: [], MemberBearer: [] }]
      parameters:
        - $ref: '#/components/parameters/StoreCode'
        - $ref: '#/components/parameters/ItemId'
        - $ref: '#/components/parameters/Locale'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/UpdateCartItemRequest' }
      responses:
        '200':
          description: 已更新并重算
          content: { application/json: { schema: { $ref: '#/components/schemas/Cart' } } }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
    delete:
      tags: [Cart]
      summary: 删除当前会员购物车行
      security: [{ StoreCode: [], MemberBearer: [] }]
      parameters:
        - $ref: '#/components/parameters/StoreCode'
        - $ref: '#/components/parameters/ItemId'
        - $ref: '#/components/parameters/Locale'
        - name: expected_version
          in: query
          required: true
          schema: { type: integer, minimum: 1 }
      responses:
        '204': { description: 已删除 }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
...
components:
  securitySchemes:
    StoreCode:
      type: apiKey
      in: header
      name: X-Store-Code
      description: 受信商城业务编码；不能由请求体 store_id 替代
    MemberBearer:
      type: http
      scheme: bearer
      bearerFormat: Zalo-Shop-Member
      description: 商城绑定会员访问令牌；商城必须与 X-Store-Code 一致
    AdminBearer:
      type: http
      scheme: bearer
      bearerFormat: Zalo-Shop-Admin
      description: 完成密码与 TOTP 的商城管理员令牌
  parameters:
    StoreCode:
      name: X-Store-Code
      in: header
      required: true
      schema: { $ref: '#/components/schemas/Code' }
    Locale:
      name: locale
      in: query
      schema: { type: string, enum: [vi, zh, en], default: vi }
...
    Cursor:
      name: cursor
      in: query
      schema: { type: string, minLength: 1, maxLength: 1000 }
...
    SkuCode:
      name: skuCode
      in: path
      required: true
      schema: { $ref: '#/components/schemas/Code' }
    ItemId:
      name: itemId
      in: path
      required: true
      schema: { type: string, format: uuid }
...
    CouponCode:
      name: couponCode
      in: path
      required: true
      schema: { $ref: '#/components/schemas/Code' }
  responses:
    BadRequest:
      description: 输入或商城 Header 无效
      content: { application/json: { schema: { $ref: '#/components/schemas/ApiError' } } }
    Unauthorized:
      description: 认证缺失、过期或无效
      content: { application/json: { schema: { $ref: '#/components/schemas/ApiError' } } }
    Forbidden:
      description: 商城不匹配、未授权或缺少权限；不泄露资源存在性
      content: { application/json: { schema: { $ref: '#/components/schemas/ApiError' } } }
    NotFound:
      description: 当前商城内资源不存在
      content: { application/json: { schema: { $ref: '#/components/schemas/ApiError' } } }
    Conflict:
      description: 版本、状态、库存或幂等键冲突
      content: { application/json: { schema: { $ref: '#/components/schemas/ApiError' } } }
    RateLimited:
      description: 请求过于频繁
      headers: { Retry-After: { schema: { type: integer, minimum: 1 } } }
      content: { application/json: { schema: { $ref: '#/components/schemas/ApiError' } } }
...
  schemas:
    Code:
      type: string
      minLength: 2
      maxLength: 64
      pattern: '^[a-z][a-z0-9-]*$'
    VndAmount:
      type: integer
      format: int64
      minimum: 0
      maximum: 9007199254740991
    Quantity:
      type: integer
      minimum: 1
      maximum: 99
    Locale:
      type: string
      enum: [vi, zh, en]
    CursorPage:
      type: object
      required: [next_cursor]
      properties:
        next_cursor: { type: [string, 'null'], maxLength: 1000 }
    ApiError:
      type: object
      required: [code, correlation_id, message_key]
      properties:
        code:
          type: string
          enum:
            - AUTHENTICATION_FAILED
            - AUTHENTICATION_REQUIRED
            - AUTHORIZATION_DENIED
            - INPUT_INVALID
            - RESOURCE_NOT_FOUND
            - STORE_CONTEXT_INVALID
            - CONFLICT
            - RATE_LIMITED
            - INTERNAL_ERROR
        correlation_id: { type: string, minLength: 1, maxLength: 128 }
        message_key: { type: string, minLength: 1, maxLength: 160 }
        details:
          type: object
          description: >-
            可含稳定 reason_code（例如 VERSION_CONFLICT、IDEMPOTENCY_KEY_REUSED、
            AVAILABLE_INSUFFICIENT、COUPON_INVALID）与脱敏字段错误；不得含跨商城对象、SQL 或堆栈。
          properties:
            reason_code:
              type: string
              enum:
                - VERSION_CONFLICT
                - IDEMPOTENCY_KEY_REUSED
                - AVAILABLE_INSUFFICIENT
                - RESERVED_INSUFFICIENT
                - RESERVATION_TRANSITION_INVALID
                - QUANTITY_OVERFLOW
                - PRODUCT_UNAVAILABLE
                - SKU_UNAVAILABLE
                - COUPON_INVALID
                - COUPON_CLAIM_LIMIT
                - COUPON_MEMBER_REQUIRED
                - COUPON_STATE_CONFLICT
                - MEMBER_INELIGIBLE
                - PROMOTION_STATE_CONFLICT
                - PROMOTION_VERSION_STATE_CONFLICT
                - PROMOTION_RULE_INVALID
                - CART_LINE_CONFLICT
                - IMPORT_INVALID
...
    MemberCoupon:
      type: object
      additionalProperties: false
      required: [id, code, status, claimed_at, expires_at]
      properties:
        id: { type: string, format: uuid }
        code: { $ref: '#/components/schemas/Code' }
        status: { type: string, enum: [CLAIMED, EXPIRED, DISABLED] }
        claimed_at: { type: string, format: date-time }
        expires_at: { type: [string, 'null'], format: date-time }
    MemberCouponPage:
      unevaluatedProperties: false
      allOf:
        - $ref: '#/components/schemas/CursorPage'
        - type: object
          required: [items]
          properties:
            {
              items:
                { type: array, maxItems: 50, items: { $ref: '#/components/schemas/MemberCoupon' } },
            }
...
    AppliedPriceRule:
      type: object
      additionalProperties: false
      required: [bucket, code, version_id, basis_vnd, discount_vnd]
      properties:
        bucket: { type: string, enum: [ITEM, ORDER, COUPON] }
        code: { $ref: '#/components/schemas/Code' }
        version_id: { type: string, format: uuid }
        basis_vnd: { $ref: '#/components/schemas/VndAmount' }
        discount_vnd: { $ref: '#/components/schemas/VndAmount' }
    RejectedPriceRule:
      type: object
      additionalProperties: false
      required: [bucket, code, version_id, reason]
      properties:
        bucket: { type: string, enum: [ITEM, ORDER, COUPON, SHIPPING] }
        code: { $ref: '#/components/schemas/Code' }
        version_id: { type: string, format: uuid }
        reason:
          type: string
          enum:
            [
              NOT_STARTED,
              ENDED,
              TARGET_MISMATCH,
              MINIMUM_NOT_MET,
              MINIMUM_QUANTITY_NOT_MET,
              NOT_BEST,
              NOT_STACKABLE,
              MEMBER_INELIGIBLE,
              COUPON_INVALID,
            ]
    PricingQuoteLine:
      type: object
      additionalProperties: false
      required:
        [
          sku_code,
          quantity,
          base_unit_price_vnd,
          base_subtotal_vnd,
          discount_vnd,
          payable_vnd,
          issues,
        ]
      properties:
        sku_code: { $ref: '#/components/schemas/Code' }
        quantity: { $ref: '#/components/schemas/Quantity' }
        base_unit_price_vnd: { $ref: '#/components/schemas/VndAmount' }
        base_subtotal_vnd: { $ref: '#/components/schemas/VndAmount' }
        discount_vnd: { $ref: '#/components/schemas/VndAmount' }
        payable_vnd: { $ref: '#/components/schemas/VndAmount' }
        applied_rules: { type: array, items: { $ref: '#/components/schemas/AppliedPriceRule' } }
        rejected_rules: { type: array, items: { $ref: '#/components/schemas/RejectedPriceRule' } }
        issues:
          type: array
          items:
            {
              type: string,
              enum: [PRODUCT_UNAVAILABLE, SKU_UNAVAILABLE, OUT_OF_STOCK, STOCK_INSUFFICIENT],
            }
    ShippingQualification:
      type: object
      additionalProperties: false
      required: [status, candidates]
      properties:
        status: { type: string, enum: [NOT_REQUESTED, ELIGIBLE_PENDING_FREIGHT, NOT_ELIGIBLE] }
        candidates:
          type: array
          maxItems: 1
          items:
            type: object
            additionalProperties: false
            required: [code, version_id]
            properties:
              code: { $ref: '#/components/schemas/Code' }
              version_id: { type: string, format: uuid }
    PricingQuote:
      type: object
      additionalProperties: false
      required:
        [
          currency,
          quoted_at,
          quote_hash,
          lines,
          base_subtotal_vnd,
          discount_vnd,
          merchandise_payable_vnd,
          order_payable_vnd,
          applied_rules,
          rejected_rules,
          shipping_qualification,
        ]
      properties:
        currency: { type: string, const: VND }
        quoted_at: { type: string, format: date-time }
        quote_hash: { type: string, pattern: '^[a-f0-9]{64}$' }
        lines:
          {
            type: array,
            minItems: 1,
            maxItems: 100,
            items: { $ref: '#/components/schemas/PricingQuoteLine' },
          }
        base_subtotal_vnd: { $ref: '#/components/schemas/VndAmount' }
        discount_vnd: { $ref: '#/components/schemas/VndAmount' }
        merchandise_payable_vnd: { $ref: '#/components/schemas/VndAmount' }
        order_payable_vnd:
          type: 'null'
          description: M3 没有地址/运费事实，最终订单应付必须为 null
        applied_rules: { type: array, items: { $ref: '#/components/schemas/AppliedPriceRule' } }
        rejected_rules: { type: array, items: { $ref: '#/components/schemas/RejectedPriceRule' } }
        shipping_qualification: { $ref: '#/components/schemas/ShippingQualification' }
    SetCartItemRequest:
      type: object
      additionalProperties: false
      required: [quantity]
      properties:
        quantity: { $ref: '#/components/schemas/Quantity' }
        selected: { type: boolean, default: true }
    UpdateCartItemRequest:
      type: object
      additionalProperties: false
      required: [expected_version]
      minProperties: 2
      properties:
        expected_version: { type: integer, minimum: 1 }
        quantity: { $ref: '#/components/schemas/Quantity' }
        selected: { type: boolean }
        replacement_sku_code: { $ref: '#/components/schemas/Code' }
    CartIssue:
      type: object
      additionalProperties: false
      required: [code, blocking]
      properties:
        code:
          {
            type: string,
            enum:
              [
                PRODUCT_UNAVAILABLE,
                SKU_UNAVAILABLE,
                OUT_OF_STOCK,
                STOCK_INSUFFICIENT,
                PRICE_CHANGED,
                PROMOTION_CHANGED,
              ],
          }
        blocking: { type: boolean }
    CartMedia:
      type: object
      additionalProperties: false
      required: [alt_text, expires_at, height, url, width]
      properties:
        alt_text: { type: string, maxLength: 500 }
        expires_at: { type: string, format: date-time }
        height: { type: [integer, 'null'], minimum: 1 }
        url: { type: string, format: uri }
        width: { type: [integer, 'null'], minimum: 1 }
    CartItem:
      type: object
      additionalProperties: false
      required:
        [
          id,
          version,
          sku_code,
          quantity,
          selected,
          added_unit_price_vnd,
          current_unit_price_vnd,
          current_subtotal_vnd,
          available_quantity,
          issues,
        ]
      properties:
        id: { type: string, format: uuid }
        version: { type: integer, minimum: 1 }
        sku_code: { $ref: '#/components/schemas/Code' }
        quantity: { $ref: '#/components/schemas/Quantity' }
        selected: { type: boolean }
        added_unit_price_vnd: { $ref: '#/components/schemas/VndAmount' }
        current_unit_price_vnd: { $ref: '#/components/schemas/VndAmount' }
        current_subtotal_vnd: { $ref: '#/components/schemas/VndAmount' }
        available_quantity: { type: integer, minimum: 0, maximum: 2147483647 }
        issues: { type: array, items: { $ref: '#/components/schemas/CartIssue' } }
        product:
          type: object
          additionalProperties: false
          required: [code, name, requested_locale, resolved_locale, primary_media, available_skus]
          properties:
            code: { $ref: '#/components/schemas/Code' }
            name: { type: string, minLength: 1, maxLength: 240 }
            requested_locale: { $ref: '#/components/schemas/Locale' }
            resolved_locale: { $ref: '#/components/schemas/Locale' }
            primary_media:
              oneOf:
                - $ref: '#/components/schemas/CartMedia'
                - type: 'null'
            available_skus:
              type: array
              maxItems: 100
              items:
                type: object
                additionalProperties: false
                required: [code, option_values]
                properties:
                  code: { $ref: '#/components/schemas/Code' }
                  option_values:
                    type: array
                    maxItems: 20
                    items: { type: object, additionalProperties: true }
        sku:
          type: object
          additionalProperties: false
          required: [code, media, option_values]
          properties:
            code: { $ref: '#/components/schemas/Code' }
            media:
              oneOf:
                - $ref: '#/components/schemas/CartMedia'
                - type: 'null'
            option_values:
              type: array
              maxItems: 20
              items: { type: object, additionalProperties: true }
    Cart:
      type: object
      additionalProperties: false
      required: [id, version, items, blocking, quote]
      properties:
        id: { type: string, format: uuid }
        version: { type: integer, minimum: 1 }
        items: { type: array, items: { $ref: '#/components/schemas/CartItem' } }
        blocking: { type: boolean }
        quote:
          oneOf: [{ $ref: '#/components/schemas/PricingQuote' }, { type: 'null' }]
          description: 仅选中且无阻断问题的行参与；空车或全部阻断时为 null
...